+++ modules/service/src/test/scala/lucuma/odb/graphql/mutation/createProgram.scala
@@ -259,18 +259,3 @@ class createProgram extends OdbSuite {
   }
 
 }
-
-trait CreateProgramOps { this: OdbSuite =>
-
-  def createProgramAs(user: User): IO[Program.Id] =
-    query(user, "mutation { createProgram(input: { SET: { name: null } }) { program { id } } }").flatMap { js =>
-      js.hcursor
-        .downField("createProgram")
-        .downField("program")
-        .downField("id")
-        .as[Program.Id]
-        .leftMap(f => new RuntimeException(f.message))
-        .liftTo[IO]
-    }
-
-}
